+++ definitions/liquipedia_db.lua
